Introduce NodeAllocatorService
